fixed theme-router again
last commit to theme-router fixed the case for planets hosted on root
level, but broke for the ones hosted on a subdir. fortunately, this
commit should fix both cases. basically, we cannot redirect to
//some-path/resource, that should be /some-path/resource instead. the
problem is caused when the the script path is '/' so dirname() returns
the root dir '/'. For all the other cases it returns
'somedir/somesubdir' without the leading slash.